requirements.txt
openai==1.101.0
azure-search-documents==11.4.0
pydantic-settings==2.9.1
python-dotenv==1.1.1
fastapi==0.116.1
python-pptx==1.0.2
pdf2image==1.17.0
pillow==11.3.0
docx2pdf==0.1.8
python-docx==1.1.2
python-pptx==1.0.2
PyMuPDF==1.25.5
uvicorn[standard]==0.24.0